Reworks tests to allow for weaker access
Android tests, or tests that run on real device, run in a different package from the test runner.  This caused a lot of issues when using Mockito to mock dependencies that weren't public.  This led to making a lot of things public that really shouldn't be.  This diff moves some tests to /test and makes them pure unit tests.  Pure unit tests run on a dev computers JVM and are in the same package as the code under test.  Package private deps can be mocked this way.  I only had to rework a little bit of actual source code.  Most of these changes reduce method and class visibility only.

Also, adds copyright everywhere and adds it to Intellij for the project so the copyright will automatically be added to new classes.

Also, updates the testAllModules task to run the new unit tests in addition to the Android tests.