Tweaks datafile loading pipeline to ensure callback is hit
Currently it's possible for OptimizelyStartListener#onStart(OptimizelyClient) method to never get called if datafile loading fails from disk and the CDN.  We should still call it with the dummy instance.  It's highly likely that developers will gate their app depending on whether or not Optimizely loaded and calls the callback.  For example, they could start Optimizely in a splash screen activity and launch the next activity once Optimizely is loaded.