Update proguard-rules.pro
The java core needs this class names to be intact to detect the right parser to use.  The Android SDK compiles gson so it should be available.  Ebay's proguard issue indicated that org.json.JSONObject was being used, however.  The reason gson is compiled is because Android's JSONObject and the JDKs are different and incompatible.  It's difficult to override Android's because it's part of the library.